cloud-native_ai/3-crud/crud_environment_variables/crud/pyproject.toml
[project]
name = "crud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-env>=1.0.0",
    "uvicorn>=0.40.0",
]

[tool.uv.workspace]
members = [
    "CRUD",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
